Fixed the state store issue
In Gobblin MR, if a job fails even before the Hadoop MR job starts, no task states will be collected and we should not persist job state in this case. The method persistJobState() in AbstractJobLauncher actually takes this into consideration by not persisting job state if the job has not completed. However, in launchJob(), when exceptions thrown by runJob is caught, the job state is set to FAILED, preventing the above guard from taking effect. This commit fixes it.

Signed-off-by: Yinan Li <ynli@linkedin.com>

RB=372621
R=etl-infra-reviewers
A=stakiar