Fixed unit tests around JobLauncher implementations after recent changes on writer file naming
A recent change on writer file naming allows user to fully control the writer file name including its extension through the configuration property writer.file.name. This property used to be used as the prefix for writer file names that will also include the task IDs, which guarantees that each task writes to a different file. However, the test job configurations explicitly set writer.file.name=gobblin-test, which means every tasks of a job will write to the same file. This issue shows up in LocalJobLauncherTest and MRJobLauncherTest in the form of WARN messages that complained that a writer of a task is trying to write to an existing file named gobblin-test. This commit fixed the tests by removing writer.file.name from the test job configurations.

Signed-off-by: Yinan Li <liyinan926@gmail.com>