Do not override `MultiWorkUnit` grouping in Gobblin Yarn
Currently, the `GobblinHelixJobLauncher` flattens `MultiWorkUnit`s and hope for Helix to do a balanced assignments of tasks to containers. However, both Gobblin Kafka and IDPC HDFS Purger already have their own logic of grouping `WorkUnit`s into `MultiWorkUnit`s. It's good to keep their logic and not to override it.

Signed-off-by: Yinan Li <liyinan926@gmail.com>