More changes toward an end-to-end flow with the new APIS (#1178)
Changes to the new API:

GobblinInstanceLauncher/GobblinInstanceDriver

Renamed GobblinInstanceLauncher.getName() to GobblinInstanceLauncher.getInstanceName()
Made GobblinInstanceLauncher extend configurable since it is responsible of holding the system configuration
JobCatalog/JobSpec:

Refactored the register/unregister methods into a separate interface - JobCatalogListenersContainer
Reworked JobSpec.Builder to be convention-based and be able to reuse some of the properties in the Job configuration
JobExecutionLauncher/JobExecutionDriver

Added the class JobExecutionDriver to control/monitor the execution of a single gobblin job
JobExecutionLauncher is not a factory for JobExecutionDriver instances
Add JobExecutionStateListenerContainer interface for managing JobExecutionStateListener instances
JobExecutionStatus/JobExecutionState

JobExecutionStatus is now an interface
JobExecutionState is now responsible for the state management (including callbacks)
Implementations:

Added a sample implementation JobLauncherExecutionDriver of JobExecutionDriver which acts as an adapter to the legacy JobLauncher/JobLauncherFactory API
Added DefaultGobblinInstanceDriverImpl which accepts all components but it is not responsible for managing them. This implementation can be used as a shell to implement a typical Gobblin job lifecycle
Added StandardGobblinInstanceDriver which will be responsible for instantiating and managing the instance components and passing them to DefaultGobblinInstanceDriverImpl
Added StandardGobblinInstanceLauncher-- a very rudimentary implementation to instantiate StandardGobblinInstanceDriver
Standard implementations for JobCatalogListenersContainer (JobCatalogListenersList) and JobExecutionStateListenerContainer (JobExecutionStateListeners)
The above use a generic callback infrastructure in gobblin-utility/gobblin/util/callbacks
Callbacks a typically represented through a Function where L is the listener type and R is the callback result (or Void)
There is a helper class Callback to let adding some human-readable description of the callback
Callbacks are created by a CallbackFactory which is responsble for binding any parameters that need to be passed to the callback
Callbacks are managed by CallbacksDispatcher
Changes to legacy code:

Documentation and helper methods in JobState to decrease reliance on direct access to job properties
Added a description of current state diagram for RunningState based on code inspection
Helper methods in JobContext
Added a method JobLauncherFactory. newJobLauncher() method to allow instantiating a specific JobLauncher type instead of relying on the job config